Improved WPCS compliance
