+++ lib/api/wp-customize/functions.php
@@ -5,20 +5,20 @@
  *
  * @package Beans\Framework\API\WP_Customize
  *
- * @since 1.0.0
+ * @since   1.0.0
  */
 
 /**
  * Register WP Customize Options.
  *
- * This function should only be invoked through the 'customize_register' action.
+ * This function should only be called with the 'customize_register' action.
  *
  * @since 1.0.0
  *
  * @param array  $fields {
  *      Array of fields to register.
  *
- *      @type string $id          A unique id used for the field. This id will also be used to save the value in
+ *      @type string $id          A unique ID used for the field. This ID will also be used to save the value in
  *                                the database.
  *      @type string $type        The type of field to use. Please refer to the Beans core field types for more
  *                                information. Custom field types are accepted here.
@@ -28,9 +28,9 @@
  *      @type array  $attributes  An array of attributes to add to the field. The array key defines the
  *                                attribute name and the array value defines the attribute value. Default array.
  *      @type mixed  $default     The default field value. Default false.
- *      @type array  $fields      Must only be used for 'group' field type. The array arguments are similary to the
+ *      @type array  $fields      Must only be used for the 'group' field type. The array arguments are similar to the
  *                                {@see beans_register_fields()} $fields arguments.
- *      @type bool   $db_group    Must only be used for 'group' field type. It defines whether the group of fields
+ *      @type bool   $db_group    Must only be used for the 'group' field type. It defines whether the group of fields
  *                                registered should be saved as a group in the database or as individual
  *                                entries. Default false.
  * }
@@ -72,7 +72,7 @@ function beans_register_wp_customize_options( array $fields, $section, $args = a
 	}
 
 	// Load the class only if this function is called to prevent unnecessary memory usage.
-	require_once BEANS_API_PATH . 'wp-customize/class.php';
+	require_once BEANS_API_PATH . 'wp-customize/class-beans-wp-customize.php';
 
 	new _Beans_WP_Customize( $section, $args );
 }
